requirements.txt
tensorflow-macos
tensorflow-metal
mlflow
matplotlib
seaborn
numpy
scikit-learn